+++ Sources/Mistica/Components/Cards/HighlightedCard.swift
@@ -10,7 +10,6 @@ import UIKit
 
 public class HighlightedCard: UIView {
     private enum Constants {
-        static let cornerRadius: CGFloat = 8
         static let closeButtonAlpha: CGFloat = 0.7
         static let closeButtonHeight: CGFloat = 48
         static let minHeight: CGFloat = 100
@@ -31,7 +30,6 @@ public class HighlightedCard: UIView {
         case secondary
     }
 
-    private lazy var cardAccessibilityElement = UIAccessibilityElement(accessibilityContainer: self)
     private lazy var verticalStackView = UIStackView()
     private lazy var horizontalStackView = UIStackView()
 
@@ -143,12 +141,8 @@ public class HighlightedCard: UIView {
 
     override public var accessibilityElements: [Any]? {
         get {
-            updateAccessibilityLabel()
-            // We must set the frame and be sure it is already calculated.
-            cardAccessibilityElement.accessibilityFrameInContainerSpace = bounds
-            return [
-                cardAccessibilityElement,
-                actionButton.isHidden ? nil : actionButton,
+            [
+                horizontalStackView,
                 closeButton.isHidden ? nil : closeButton
             ].compactMap { $0 }
         }
@@ -313,15 +307,6 @@ public extension HighlightedCard {
             backgroundImageView.accessibilityIdentifier = newValue
         }
     }
-
-    override var accessibilityTraits: UIAccessibilityTraits {
-        get {
-            cardAccessibilityElement.accessibilityTraits
-        }
-        set {
-            cardAccessibilityElement.accessibilityTraits = newValue
-        }
-    }
 }
 
 // MARK: Private
@@ -408,7 +393,8 @@ private extension HighlightedCard {
         closeButton.isHidden = true
 
         addBorder(color: .border)
-        makeRounded(cornerRadius: Constants.cornerRadius)
+        setMisticaRadius(.container)
+        clipsToBounds = true
     }
 
     func updateRightImageViewVisibility() {
@@ -472,12 +458,5 @@ private extension HighlightedCard {
 
     @objc func actionButtonTapped() {
         actionButtonCallback?()
-    }
-
-    func updateAccessibilityLabel() {
-        cardAccessibilityElement.accessibilityLabel = [
-            titleAccessibilityLabel,
-            subtitleAccessibilityLabel
-        ].compactMap { $0 }.joined(separator: " ")
     }
 }