+++ Sources/Mistica/Components/Stepper/SegmentView.swift
@@ -46,8 +46,10 @@ class SegmentView: UIView {
 
     override func layoutSubviews() {
         super.layoutSubviews()
-        backgroundView.makeRounded(cornerRadius: Constants.height / 2)
-        trackView.makeRounded(cornerRadius: Constants.height / 2)
+        backgroundView.setMisticaRadius(.bar)
+        trackView.setMisticaRadius(.bar)
+        backgroundView.clipsToBounds = true
+        trackView.clipsToBounds = true
         updateValueConstraint()
     }
 }