+++ Sources/MisticaCommon/Radius/CornerRadiusStyle.swift
@@ -1,0 +1,23 @@
+//
+//  CornerRadiusStyle.swift
+//
+//  Made with ❤️ by Novum
+//
+//  Copyright © Telefonica. All rights reserved.
+//
+
+import Foundation
+
+@frozen
+@objc public enum CornerRadiusStyle: Int, CaseIterable {
+    case avatar
+    case bar
+    case button
+    case checkbox
+    case container
+    case indicator
+    case input
+    case legacyDisplay
+    case popup
+    case sheet
+}