+++ Sources/MisticaSwiftUI/Components/Button/MisticaButton.swift
@@ -17,7 +17,6 @@ struct MisticaButton: View {
         static let minWidth: CGFloat = 76
         static let height: CGFloat = 24
         static let smallHeight: CGFloat = 16
-        static let cornerRadius: CGFloat = 4
         static let smallVerticalPadding: CGFloat = 8
         static let verticalPadding: CGFloat = 12
         static let horizontalPadding: CGFloat = 16
@@ -91,7 +90,7 @@ struct MisticaButton: View {
         .padding(.horizontal, Constants.horizontalPadding)
         .background(backgroundColor)
         .foregroundColor(foregroundColor)
-        .border(cornerRadius: Constants.cornerRadius, borderColor: borderColor)
+        .border(radiusStyle: .button, borderColor: borderColor)
         .padding(EdgeInsets(top: 0, leading: leadingInset, bottom: 0, trailing: trailingInset))
         .allowsHitTesting(!loadingInfo.isLoading)
     }