+++ Tests/MisticaTests/UI/FeedbackTests.swift
@@ -71,6 +71,19 @@ final class FeedbackTests: XCTestCase {
         )
     }
 
+    func testFeedbackTitleAndSubtitle() {
+        assertSnapshotForAllBrandsAndStyles(
+            as: .image,
+            viewBuilder: feedbackViewController(
+                style: .feedback(icon: UIImage(systemName: "swift")!),
+                title: Constants.singleLineTitle,
+                subtitle: Constants.singleLineSubtitle,
+                primaryAction: .none,
+                secondaryAction: .none
+            )
+        )
+    }
+
     func testErrorTitleSubtitleAndErrorReference() {
         assertSnapshotForAllBrandsAndStyles(
             as: .image,